Jaeger tracing support
This introduces utility code in csi-common which can be used to
initialize a global OpenTracing tracer and then log information both
in the normal log and the trace. The hostpath-driver uses this.

Right now, the global tracer is always the one from Jaeger. Other
tracing implementations could be added by changing only the InitTracer
implementation.